Row_printKBytes use unsigned int i

+++ Row.c
@@ -232,7 +232,7 @@ void Row_printKBytes(RichString* str, unsigned long long number, bool coloring) 
    const size_t maxUnitIndex = (sizeof(number) * CHAR_BIT - 1) / 10 + 1;
    const bool canOverflow = maxUnitIndex >= ARRAYSIZE(unitPrefixes);
 
-   uint8_t i = 1;
+   unsigned int i = 1;
    int prevUnitColor;
    // Convert KiB to (1/100) of MiB
    unsigned long long hundredths = (number / 256) * 25 + (number % 256) * 25 / 256;
